requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
aiosqlite==0.20.0
httpx==0.28.1
itsdangerous==2.2.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic-settings==2.7.1
python-multipart==0.0.20
